scripts/requirements.txt
notion-client==2.2.1
python-slugify==8.0.4
requests==2.31.0
